requirements.txt
# local package
-e .

# external requirements
dvc
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
tensorflow
numpy
pandas
matplotlib
flask
pickle
opencv-python
matplotlib